+++ lib/page/marker/report_page.dart
@@ -6,9 +6,9 @@ import 'package:insigno_frontend/networking/authentication.dart';
 import 'package:insigno_frontend/networking/backend.dart';
 import 'package:insigno_frontend/networking/data/map_marker.dart';
 import 'package:insigno_frontend/networking/data/marker_type.dart';
-import 'package:insigno_frontend/page/map/location_provider.dart';
+import 'package:insigno_frontend/provider/location_provider.dart';
 import 'package:insigno_frontend/page/marker/add_images_widget.dart';
-import 'package:insigno_frontend/page/user/auth_user_provider.dart';
+import 'package:insigno_frontend/provider/auth_user_provider.dart';
 import 'package:insigno_frontend/util/error_messages.dart';
 import 'package:insigno_frontend/util/error_text.dart';
 import 'package:insigno_frontend/util/pair.dart';
