+++ lib/page/marker/resolve_page.dart
@@ -5,9 +5,9 @@ import 'package:get_it_mixin/get_it_mixin.dart';
 import 'package:insigno_frontend/networking/authentication.dart';
 import 'package:insigno_frontend/networking/backend.dart';
 import 'package:insigno_frontend/networking/data/map_marker.dart';
-import 'package:insigno_frontend/page/map/location_provider.dart';
+import 'package:insigno_frontend/provider/location_provider.dart';
 import 'package:insigno_frontend/page/marker/add_images_widget.dart';
-import 'package:insigno_frontend/page/user/auth_user_provider.dart';
+import 'package:insigno_frontend/provider/auth_user_provider.dart';
 import 'package:insigno_frontend/util/error_messages.dart';
 import 'package:insigno_frontend/util/error_text.dart';
 import 'package:insigno_frontend/util/nullable.dart';
